fix mindswap
